+++ resources/views/components/copy_button.blade.php
@@ -1,0 +1,6 @@
+<div class="pb-2">
+    <a href="#" class="js-copy" data-message="{{ trans('general.copied_to_clipboard') }}"
+        data-url="{{ $text }}" data-toggle="tooltip" data-trigger="manual" data-container="body"
+        title="{{ trans('general.copy') }}"><i class="fa fa-clipboard" aria-hidden="true"></i>
+        {!! trans('general.copy') !!}</a>
+</div>