+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.67.0] - 2022-10-04
+### Fixed
+- lazy loading items for reindex
+
+### Added
+- artwork info to citation note on download
+- import iip images command
+
 ## [2.66.0] - 2022-09-22
 ### Fixed
 - item detail view to eliminate N+1 queries